yolo5/requirements.txt
flask>=2.3.2
pyyaml
loguru
boto3
pymongo
